// src/features/gestion-cobranzas/modules/gestion-deudor/api/deudoresGestionDeudorApi.types.ts
export interface DeudorGestionDeudorApi {
  nId_PersDeudor: unknown;
  nro: unknown;
  zonaCampanna: unknown;
  nId_Cliente: unknown;
  nId_Contrato: unknown;
  nId_Cartera: unknown;
  cartera: unknown;
  codigoCliente: unknown;
  deudor: unknown;
  importe: unknown;
  saldo: unknown;
  fechaUltimaGestionCALL: unknown;
  ultimaGestionCALL: unknown;
  cantidadGestionCALL: unknown;
  fechaUltimaGestionCAMPO: unknown;
  ultimaGestionCAMPO: unknown;
  cantidadGestionCAMPO: unknown;
  fechaPromesa: unknown;
  mejorStatus: unknown;
}